backend/app/schemas/auth.py
from pydantic import BaseModel


class AuthTokenRequest(BaseModel):
    id_token: str


class AuthTokenResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"
    user_id: str
    is_new_user: bool
