backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
psycopg[binary]==3.2.9
alembic==1.16.4
pydantic-settings==2.10.1
python-dotenv==1.1.1
redis==5.2.1
celery==5.5.3
firebase-admin==6.9.0
google-cloud-storage==2.19.0
httpx==0.28.1
meilisearch==0.34.0
python-multipart==0.0.20
orjson==3.11.2
